requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
Click==7.0
cycler==0.10.0
future==0.18.2
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
opencv-python==4.1.2.30
pathspec==0.6.0
pkg-resources==0.0.0
pyglet==1.4.8
pyparsing==2.4.5
pyserial==3.4
python-dateutil==2.8.1
PyYAML==5.1.2
regex==2019.11.1
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
